[Fix] Device inference
