+++ pyproject.toml
@@ -16,7 +16,7 @@ authors = [
 ]
 requires-python = ">=3.8,<3.12"
 license = {text = "Apache 2.0"}
-version = "1.0.3"
+version = "1.0.4"
 classifiers=[
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python",
@@ -28,14 +28,10 @@ classifiers=[
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
 ]
-dependencies = [
-    "numpy",
-    "torch",
-]
+dependencies = ["torch"]
 
 [project.optional-dependencies]
-dev = ["black", "pytest", "pytest-xdist", "pytest-cov", "flake8", "mypy", "pre-commit", "ruff", "nbconvert",
-  "ipykernel", "matplotlib"]
+dev = ["black", "pytest", "pytest-xdist", "pytest-cov", "flake8", "mypy", "pre-commit", "ruff", "nbconvert", "matplotlib"]
 
 [tool.hatch.envs.tests]
 features = [
