+++ pyqtorch/circuit.py
@@ -2,27 +2,25 @@ from __future__ import annotations
 
 from typing import Any, Iterator
 
-import torch
+from torch import Tensor, device
+from torch.nn import Module, ModuleList
 
-from pyqtorch.abstract import AbstractOperator
 from pyqtorch.utils import DiffMode, State, overlap, zero_state
 
 
-class QuantumCircuit(torch.nn.Module):
-    def __init__(
-        self, n_qubits: int, operations: list[AbstractOperator], diff_mode: DiffMode = DiffMode.AD
-    ):
+class QuantumCircuit(Module):
+    def __init__(self, n_qubits: int, operations: list[Module], diff_mode: DiffMode = DiffMode.AD):
         super().__init__()
         self.n_qubits = n_qubits
-        self.operations = torch.nn.ModuleList(operations)
+        self.operations = ModuleList(operations)
         self.diff_mode = diff_mode
 
-    def __mul__(self, other: AbstractOperator | QuantumCircuit) -> QuantumCircuit:
+    def __mul__(self, other: Module | QuantumCircuit) -> QuantumCircuit:
         n_qubits = max(self.n_qubits, other.n_qubits)
         if isinstance(other, QuantumCircuit):
             return QuantumCircuit(n_qubits, self.operations.extend(other.operations))
 
-        elif isinstance(other, AbstractOperator):
+        elif isinstance(other, Module):
             return QuantumCircuit(n_qubits, self.operations.append(other))
 
         else:
@@ -43,22 +41,22 @@ class QuantumCircuit(torch.nn.Module):
     def __hash__(self) -> int:
         return hash(self.__key())
 
-    def run(self, state: State = None, values: dict[str, torch.Tensor] = {}) -> State:
+    def run(self, state: State = None, values: dict[str, Tensor] = {}) -> State:
         if state is None:
             state = self.init_state()
         for op in self.operations:
             state = op(state, values)
         return state
 
-    def forward(self, state: State, values: dict[str, torch.Tensor] = {}) -> State:
+    def forward(self, state: State, values: dict[str, Tensor] = {}) -> State:
         return self.run(state, values)
 
     def expectation(
         self,
-        values: dict[str, torch.Tensor],
+        values: dict[str, Tensor],
         observable: QuantumCircuit,
         state: State = None,
-    ) -> torch.Tensor:
+    ) -> Tensor:
         if observable is None:
             raise ValueError("Please provide an observable to compute expectation.")
         if state is None:
@@ -74,15 +72,19 @@ class QuantumCircuit(torch.nn.Module):
             )
 
     @property
-    def _device(self) -> torch.device:
+    def _device(self) -> device:
         try:
             (_, buffer) = next(self.named_buffers())
             return buffer.device
         except StopIteration:
-            return torch.device("cpu")
+            return device("cpu")
 
-    def init_state(self, batch_size: int = 1) -> torch.Tensor:
+    def init_state(self, batch_size: int = 1) -> Tensor:
         return zero_state(self.n_qubits, batch_size, device=self._device)
 
     def reverse(self) -> QuantumCircuit:
-        return QuantumCircuit(self.n_qubits, torch.nn.ModuleList(list(reversed(self.operations))))
+        return QuantumCircuit(self.n_qubits, ModuleList(list(reversed(self.operations))))
+
+    def to(self, device: device) -> QuantumCircuit:
+        self.operations = ModuleList([op.to(device) for op in self.operations])
+        return self
